refactor(T11455): decouple domain-handler subgraph from cleo-internals (R3-K1 · SG-RUNTIME-UNIFICATION)
Relocate the shared handler dependencies that coupled the 27 domain handlers
in packages/cleo/src/dispatch/domains/ to cleo-internal engines/ / lib/ /
cli/paths / nexus-decorator into @cleocode/runtime/gateway, mirroring the
R3-T3 (T11447) dispatcher/registry/meta relocation:

  - lib/engine.ts barrel        -> @cleocode/runtime/gateway/engine.ts
  - engines/_error.ts           -> @cleocode/runtime/gateway/engine-error.ts
  - lib/exit-codes.ts           -> @cleocode/runtime/gateway/exit-codes.ts
  - engines/template-parser.ts  -> @cleocode/runtime/gateway/template-parser.ts
  - nexus-decorator.ts          -> @cleocode/runtime/gateway/nexus-decorator.ts
  - lib/job-manager-accessor.ts -> @cleocode/runtime/gateway/job-manager-accessor.ts
  - lib/background-jobs.ts       -> @cleocode/runtime/gateway/background-jobs.ts
  - cli/paths.ts constants       -> @cleocode/runtime/gateway/paths.ts (3 consumed)

Domain handlers now import only from @cleocode/runtime/@cleocode/core/
@cleocode/contracts. Thin re-export shims remain at every old cleo-internal
path (zero behavior change). @cleocode/runtime gains a drizzle-orm dep (used by
the relocated background-jobs) and exposes the relocated surface via the
gateway subpath; it still has NO @cleocode/cleo dependency, so the runtime can
assemble the handler map without importing cleo — unblocking the R3 transport
adapters (T11448 MCP, T11449 CLI-RPC, T11450 Studio HTTP+SSE).
