+++ packages/runtime/src/gateway/engine.ts
@@ -1,0 +1,310 @@
+/**
+ * Engine Re-exports for Gateway Domain Handlers
+ *
+ * Centralizes the engine imports consumed by the domain handlers into a single
+ * barrel. Every symbol re-exported here ultimately resolves to
+ * `@cleocode/core/internal` (the canonical business-logic SSoT) or core's
+ * template parser — this file adds NO logic of its own.
+ *
+ * @remarks
+ * Relocated from `packages/cleo/src/dispatch/lib/engine.ts` into
+ * `@cleocode/runtime/gateway` (R3-K1 · T11455 · SG-RUNTIME-UNIFICATION) so the
+ * runtime can assemble the handler map without importing `@cleocode/cleo`.
+ * Mirrors the `meta` / `registry` / `dispatcher` relocations (R3-T3 · T11447).
+ * A thin re-export shim remains at the old `lib/engine.ts` path so the
+ * in-package import sites compile unchanged (zero behavior change).
+ *
+ * @epic T4820
+ * @task T4815
+ * @task T11455
+ */
+
+export type { MinimalTaskRecord, Session as SessionRecord, TaskRecord } from '@cleocode/contracts';
+export type { OrchestratePlanResult } from '@cleocode/contracts/operations/orchestrate';
+// Task operations — re-exported from @cleocode/core/internal (T1568 / ADR-057 / ADR-058)
+// task-engine.ts was deleted in Wave 5; all symbols now live in core/tasks/.
+// Orchestrate — types and functions from @cleocode/core/internal (T1570 ADR-057)
+// System — type aliases for backward compat (T1571: system-engine.ts deleted, canonical sources = @cleocode/core/internal)
+export type {
+  ArchiveStatsResult as ArchiveStatsData,
+  AuditResult as AuditData,
+  BackupEntry as BackupEntryData,
+  BackupResult as BackupData,
+  CleanupResult as CleanupData,
+  CompactTask,
+  ComplianceData,
+  ContextData,
+  DashboardData,
+  DiagnosticsResult as DiagnosticsData,
+  FileRestoreResult,
+  HealthResult as HealthData,
+  HelpData,
+  InjectGenerateResult as InjectGenerateData,
+  IvtrHistoryEntry,
+  LabelsResult as LabelsData,
+  LifecycleStageEntry,
+  LogQueryData,
+  MigrateResult as MigrateData,
+  OrchestratePlanInput,
+  PathsData,
+  PlanWarning,
+  PlanWave,
+  PlanWorkerEntry,
+  RestoreResult as RestoreData,
+  RuntimeDiagnostics as RuntimeData,
+  SafestopResult as SafestopData,
+  ScaffoldHubData,
+  StatsData,
+  SyncData,
+  SystemMetricsResult as MetricsData,
+  UncancelResult as UncancelData,
+} from '@cleocode/core/internal';
+// Pipeline manifest functions (moved from memory domain in T5241)
+// Release engine (T1572 — ENG-MIG-5: migrated to @cleocode/core/internal)
+// Session engine (T1573 — ENG-MIG-6: migrated to @cleocode/core/internal)
+// Lifecycle engine ops — re-exported from @cleocode/core/internal (T1576 / ENG-MIG-9 / ADR-057)
+// lifecycle-engine.ts shrunk to a shim; all symbols now live in core/lifecycle/engine-ops.ts.
+// Hooks engine (ENG-MIG-12 / T1579 — logic in @cleocode/core/internal)
+// Init engine-ops — migrated to @cleocode/core/internal (ENG-MIG-14 / T1581 / ADR-057)
+// Config engine-ops — migrated to @cleocode/core/internal (ENG-MIG-15 / T1582 / ADR-057)
+// config-engine.ts shrunk to a shim; all symbols now live in core/config/engine-ops.ts.
+// Codebase map — re-exported from core (engine shim deleted, T1584)
+// Memory engine — brain.db cognitive memory (T5241 cutover)
+// Validate engine — direct from core (was a rename shim, now inlined here)
+export {
+  addTaskWithSessionScope,
+  completeTaskStrict,
+  configGet,
+  configListPresets,
+  configSet,
+  configSetPreset,
+  coreBatchValidate as validateBatchValidate,
+  coreCoherenceCheck as validateCoherenceCheck,
+  coreComplianceRecord as validateComplianceRecord,
+  coreComplianceSummary as validateComplianceSummary,
+  coreComplianceViolations as validateComplianceViolations,
+  coreTestCoverage as validateTestCoverage,
+  coreTestRun as validateTestRun,
+  coreTestStatus as validateTestStatus,
+  coreValidateManifest as validateManifestOp,
+  coreValidateOutput as validateOutput,
+  coreValidateProtocol as validateProtocol,
+  coreValidateSchema as validateSchemaOp,
+  coreValidateTask as validateTaskOp,
+  type DecisionRecord,
+  ensureInitialized,
+  filterEntries as filterManifestEntries,
+  getVersion,
+  type HookMatrixResult,
+  initProject,
+  isAutoInitEnabled,
+  lifecycleCheck,
+  lifecycleGateFail,
+  lifecycleGatePass,
+  lifecycleGates,
+  lifecycleHistory,
+  lifecyclePrerequisites,
+  lifecycleProgress,
+  lifecycleReset,
+  lifecycleSkip,
+  lifecycleStatus,
+  type ManifestEntry as ResearchManifestEntry,
+  mapCodebase,
+  memoryBrainStats,
+  memoryContradictions,
+  memoryDecisionFind,
+  memoryDecisionStore,
+  memoryFetch,
+  memoryFind,
+  memoryGraphAdd,
+  memoryGraphContext,
+  memoryGraphNeighbors,
+  memoryGraphRelated,
+  memoryGraphRemove,
+  memoryGraphShow,
+  memoryGraphStatsFull,
+  memoryGraphTrace,
+  memoryLearningFind,
+  memoryLearningStats,
+  memoryLearningStore,
+  memoryLink,
+  memoryObserve,
+  memoryPatternFind,
+  memoryPatternStats,
+  memoryPatternStore,
+  memoryQualityReport,
+  memoryReasonSimilar,
+  memoryReasonWhy,
+  memorySearchHybrid,
+  memoryShow,
+  memorySuperseded,
+  memoryTimeline,
+  memoryUnlink,
+  orchestrateAnalyze,
+  orchestrateBootstrap,
+  orchestrateCheck,
+  orchestrateContext,
+  orchestrateCriticalPath,
+  orchestrateHandoff,
+  orchestrateNext,
+  orchestrateParallelEnd,
+  orchestrateParallelStart,
+  orchestratePlan,
+  orchestrateReady,
+  orchestrateReport,
+  orchestrateSkillInject,
+  orchestrateSpawn,
+  orchestrateSpawnExecute,
+  orchestrateStartup,
+  orchestrateStatus,
+  orchestrateUnblockOpportunities,
+  orchestrateValidate,
+  orchestrateWaves,
+  type ProviderMatrixEntry,
+  // Pipeline engine-ops (ENG-MIG-11 / T1578 — migrated to @cleocode/core/internal)
+  // pipeline-engine.ts shrunk to a shim; all symbols now live in core/pipeline/engine-ops.ts.
+  phaseAdvance,
+  phaseComplete,
+  phaseDelete,
+  phaseList,
+  phaseRename,
+  phaseSet,
+  phaseShow,
+  phaseStart,
+  pipelineManifestAppend,
+  pipelineManifestArchive,
+  pipelineManifestFind,
+  pipelineManifestList,
+  pipelineManifestPending,
+  pipelineManifestShow,
+  pipelineManifestStats,
+  queryCommonHooks,
+  queryHookProviders,
+  readManifestEntries,
+  releaseCancel,
+  releaseCommit,
+  releaseGateCheck,
+  releaseGatesRun,
+  releaseIvtrAutoSuggest,
+  releaseList,
+  releaseOpen,
+  releasePlan,
+  releasePrepare,
+  releasePrStatus,
+  releasePush,
+  releaseReconcileV2,
+  releaseRollback,
+  releaseRollbackFull,
+  releaseShow,
+  releaseTag,
+  sessionAdopt,
+  sessionArchive,
+  sessionBriefing,
+  sessionChainShow,
+  sessionCleanup,
+  // T4959: Rich debrief + chain
+  sessionComputeDebrief,
+  sessionComputeHandoff,
+  sessionContextDrift,
+  sessionContextInject,
+  sessionDebriefShow,
+  sessionDecisionLog,
+  sessionEnd,
+  // T5119: Lightweight session discovery
+  sessionFind,
+  sessionGc,
+  sessionHandoff,
+  sessionHistory,
+  sessionList,
+  sessionRecordAssumption,
+  sessionRecordDecision,
+  sessionResume,
+  sessionShow,
+  sessionStart,
+  sessionStats,
+  sessionStatus,
+  sessionSuspend,
+  sessionSwitch,
+  systemHooksMatrix,
+  taskAnalyze,
+  taskArchive,
+  taskBatchValidate,
+  taskBlockers,
+  taskCancel,
+  taskClaim,
+  taskComplete,
+  taskComplexityEstimate,
+  taskCurrentGet,
+  taskDelete,
+  taskDepends,
+  taskDeps,
+  taskDepsCycles,
+  taskDepsOverview,
+  taskDepsTree,
+  taskDepsValidate,
+  taskExists,
+  taskExport,
+  taskFind,
+  taskHistory,
+  taskImpact,
+  taskImport,
+  taskLabelList,
+  taskLabelShow,
+  taskLint,
+  taskList,
+  taskNext,
+  taskPlan,
+  taskPromote,
+  taskRelates,
+  taskRelatesAdd,
+  taskRelatesAddBatch,
+  taskRelatesFind,
+  taskRelatesRemove,
+  taskReopen,
+  taskReorder,
+  taskReparent,
+  taskRestore,
+  taskShow,
+  taskShowIvtrHistory,
+  taskShowOperation,
+  taskShowWithHistory,
+  taskSlice,
+  taskStart,
+  taskStats,
+  taskStop,
+  taskSyncLinks,
+  taskSyncLinksRemove,
+  taskSyncReconcile,
+  tasksAddBatchOp,
+  taskTree,
+  taskUnarchive,
+  taskUnclaim,
+  taskUpdate,
+  taskWorkHistory,
+  // T9937 — validate-changelog verb (Saga T9862)
+  validateChangelog,
+  validateGateVerify,
+  // T5327 + T260: Protocol validation operations (12 protocols)
+  validateProtocolArchitectureDecision,
+  validateProtocolArtifactPublish,
+  validateProtocolConsensus,
+  validateProtocolContribution,
+  validateProtocolDecomposition,
+  validateProtocolImplementation,
+  validateProtocolProvenance,
+  validateProtocolRelease,
+  validateProtocolResearch,
+  validateProtocolSpecification,
+  validateProtocolTesting,
+  validateProtocolValidation,
+} from '@cleocode/core/internal';
+// Template parser engine
+export {
+  generateTemplateConfig,
+  getTemplateForSubcommand,
+  type IssueTemplate,
+  parseIssueTemplates,
+  type TemplateConfig,
+  type TemplateSection,
+  validateLabels,
+} from './template-parser.js';
